requirements.txt
Django==1.6.5
WebOb==1.4
WebTest==2.0.15
beautifulsoup4==4.3.2
django-extensions==1.3.10
django-nose==1.2
django-tastypie==0.11.1
django-webtest==1.7.7
factory-boy==2.4.1
ipdb==0.8
ipython==2.1.0
nose==1.3.3
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
waitress==0.8.9
coverage==3.7.1